A4A > Agency Tiers: Remove feature flag logic
