+++ client/a8c-for-agencies/components/sidebar-menu/hooks/use-main-menu-items.tsx
@@ -1,4 +1,3 @@
-import config from '@automattic/calypso-config';
 import { Badge } from '@automattic/components';
 import {
 	category,
@@ -51,7 +50,6 @@ const useMainMenuItems = ( path: string ) => {
 	const translate = useTranslate();
 
 	const agency = useSelector( getActiveAgency );
-	const isTiersRevampEnabled = config.isEnabled( 'tiers-revamp' );
 
 	const menuItems = useMemo( () => {
 		let referralItems = [] as any[];
@@ -94,19 +92,15 @@ const useMainMenuItems = ( path: string ) => {
 					menu_item: 'Automattic for Agencies / Overview',
 				},
 			},
-			...( isTiersRevampEnabled
-				? [
-						{
-							icon: starEmpty,
-							path: '/',
-							link: A4A_AGENCY_TIER_LINK,
-							title: translate( 'Agency tier' ),
-							trackEventProps: {
-								menu_item: 'Automattic for Agencies / Agency Tier',
-							},
-						},
-				  ]
-				: [] ),
+			{
+				icon: starEmpty,
+				path: '/',
+				link: A4A_AGENCY_TIER_LINK,
+				title: translate( 'Agency Tier' ),
+				trackEventProps: {
+					menu_item: 'Automattic for Agencies / Agency Tier',
+				},
+			},
 			{
 				icon: box,
 				path: '/',
@@ -236,23 +230,10 @@ const useMainMenuItems = ( path: string ) => {
 						},
 				  ]
 				: [] ),
-			...( ! isTiersRevampEnabled
-				? [
-						{
-							icon: starEmpty,
-							path: '/',
-							link: A4A_AGENCY_TIER_LINK,
-							title: translate( 'Agency Tiers' ),
-							trackEventProps: {
-								menu_item: 'Automattic for Agencies / Agency Tier',
-							},
-						},
-				  ]
-				: [] ),
 		]
 			.map( ( item ) => createItem( item, path ) )
 			.filter( ( item ) => isPathAllowed( item.link, agency ) );
-	}, [ agency, isTiersRevampEnabled, path, translate ] );
+	}, [ agency, path, translate ] );
 	return menuItems;
 };
