--- client/a8c-for-agencies/sections/agency-tier/primary/agency-tier-overview-v2/style.scss
@@ -1,1 +1,0 @@
-@import '@wordpress/dataviews/build-style/style.css';
