+++ client/a8c-for-agencies/sections/overview/sidebar/agency-tier/index.tsx
@@ -1,23 +1,13 @@
-import { isEnabled } from '@automattic/calypso-config';
-import { Card, FoldableCard } from '@automattic/components';
-import { Button } from '@wordpress/components';
-import { Icon, arrowRight } from '@wordpress/icons';
-import { clsx } from 'clsx';
 import { useTranslate } from 'i18n-calypso';
-import { A4A_AGENCY_TIER_LINK } from 'calypso/a8c-for-agencies/components/sidebar-menu/lib/constants';
 import getAgencyTierInfo from 'calypso/a8c-for-agencies/sections/agency-tier/lib/get-agency-tier-info';
 import AgencyTierProgressCard from 'calypso/a8c-for-agencies/sections/agency-tier/progress-card';
-import { preventWidows } from 'calypso/lib/formatting';
-import { useSelector, useDispatch } from 'calypso/state';
+import { useSelector } from 'calypso/state';
 import { getActiveAgency } from 'calypso/state/a8c-for-agencies/agency/selectors';
-import { recordTracksEvent } from 'calypso/state/analytics/actions';
-import AgencyTierCelebrationModal from './celebration-modal';
 
 import './style.scss';
 
 export default function OverviewSidebarAgencyTier() {
 	const translate = useTranslate();
-	const dispatch = useDispatch();
 
 	const agency = useSelector( getActiveAgency );
 
@@ -30,63 +20,11 @@ export default function OverviewSidebarAgencyTier() {
 		return null;
 	}
 
-	const isTiersRevampEnabled = isEnabled( 'tiers-revamp' );
-
-	return isTiersRevampEnabled ? (
+	return (
 		<AgencyTierProgressCard
 			currentAgencyTierId={ currentAgencyTierId }
 			influencedRevenue={ influencedRevenue ?? 0 }
 			isEarlyAccess={ !! isEarlyAccess }
 		/>
-	) : (
-		<>
-			<AgencyTierCelebrationModal
-				agencyTierInfo={ currentAgencyTierInfo }
-				currentAgencyTier={ currentAgencyTierId }
-			/>
-
-			<Card className="agency-tier__card">
-				<FoldableCard
-					className="foldable-nav"
-					header={ translate( 'Your progress' ) }
-					expanded
-					compact
-					iconSize={ 18 }
-				>
-					<div className="agency-tier__bottom-content">
-						<div
-							className={ clsx( 'agency-tier__current-agency-tier-header', {
-								'is-default': ! currentAgencyTierId,
-							} ) }
-						>
-							<span className="agency-tier__current-agency-tier-icon">
-								<img src={ currentAgencyTierInfo.logo } alt={ currentAgencyTierInfo.id } />
-							</span>
-							<span className="agency-tier__current-agency-tier-title">
-								{ preventWidows( currentAgencyTierInfo.title ) }
-							</span>
-						</div>
-						{ currentAgencyTierInfo && (
-							<div className="agency-tier__current-agency-tier-description">
-								{ currentAgencyTierInfo.description }
-							</div>
-						) }
-						<Button
-							href={ A4A_AGENCY_TIER_LINK }
-							onClick={ () =>
-								dispatch(
-									recordTracksEvent( 'calypso_a4a_agency_tier_explore_tiers_and_benefits_click', {
-										agency_tier: currentAgencyTierId,
-									} )
-								)
-							}
-						>
-							<Icon icon={ arrowRight } size={ 18 } />
-							{ translate( 'Explore Tiers and benefits' ) }
-						</Button>
-					</div>
-				</FoldableCard>
-			</Card>
-		</>
 	);
 }
