+++ config/a8c-for-agencies-production.json
@@ -42,8 +42,7 @@
 		"a4a-bd-checkout": false,
 		"a4a-vip-partner-opportunity-referrals": true,
 		"jetpack/crm-downloads": true,
-		"upgrades/redirect-payments": true,
-		"tiers-revamp": true
+		"upgrades/redirect-payments": true
 	},
 	"enable_all_sections": false,
 	"sections": {
